Initial commit (make my dotfiles public)
This marks the new start of my dotfiles. Until now, this was all in a
private repo. I wanted to share it all without worrying about
inadvertently publishing an older commit with a hardcoded secret or
something sensitive in it, so I decided to commit everything I've done
to my dotfiles over the past several years in this single initial
commit.