nvim: back up config
I'm going to try out LazyVim. I enjoy tweaking my Neovim config, but
lately adjusting my config and keeping it up-to-date has been taking
more of my time than I'd like it to. LazyVim seems to offer a nice set
of plugins and default configuration, and it's made by a plugin author I
like.

As the first step of the move, this backs up my current neovim config.

We'll see how this goes 🤞