lazygit: alias config file in setup script
LazyVim offers lazygit support out of the box, so I've started using it
and have really enjoyed it so far. This will make configuring it a
little easier, by allowing me to configure it via a dotfile. There's no
config for now, but at least I have a straightforward place to add it in
the future.