nvim: install projectionist plugin
This plugin gives me back the ability to use `<leader>a` to switch to a
file's test and vice versa. This was working in my old neovim config,
but I couldn't pinpoint exactly where the behavior was coming from (as I
didn't have projectionist installed).

I set this up to work for Ruby, JS, and TS tests, but I have only
verified it in Ruby so far.