nvim: turn off inlay hints
This approach is a bit extreme, but it's the only way I could get inlay
hints to always be off. I think some combination of LazyVim and certain
LSP servers (this is mostly an issue in JS) are enabling inlay hints
after initialization, rendering any initial setting of these options
(i.e. outside of this autocmd) ineffective.