nvim: alias to force a ruby-lsp reinstall
Every now and then, I do _something_ in my editor that causes ruby-lsp
to install an older version that is incompatible with some tooling that
my work projects use. Eventually I'll dig into the root issue, but for
now, this will give me an easy escape hatch to help get ruby-lsp back to
a working state when it breaks.